// frontend/src/api/model/documentResponse.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * MapStack API
 * AI学習プラットフォーム MapStack のバックエンドAPI
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentResponseDescription } from './documentResponseDescription';

/**
 * ドキュメント情報レスポンス
 */
export interface DocumentResponse {
  title: string;
  description?: DocumentResponseDescription;
  id: string;
  created_at: string;
  updated_at: string;
}